+++ etc/hatchet.sh
@@ -2,15 +2,16 @@
 
 set -e
 
-[ "$CIRCLE_PROJECT_USERNAME" != "heroku" ] && echo "Run tests manually for forked PRs." && exit 0
+ci_repo_owner=${CIRCLE_PROJECT_USERNAME:-${GITHUB_REPOSITORY_OWNER}}
+ci_repo_name=${CIRCLE_PROJECT_REPONAME:-${GITHUB_REPOSITORY}}
+ci_branch=${CIRCLE_BRANCH:-${GITHUB_HEAD_REF:-$GITHUB_REF_NAME}}
 
-if [[ "$CIRCLE_PROJECT_REPONAME" == "nodebin" ]]; then
+[ "$ci_repo_owner" != "heroku" ] && echo "Run tests manually for forked PRs." && exit 0
+
+if [[ "$ci_repo_name" == *nodebin ]]; then
   HATCHET_BUILDPACK_BRANCH="main"
-elif [ -n "$CIRCLE_BRANCH" ]; then
-  HATCHET_BUILDPACK_BRANCH="$CIRCLE_BRANCH"
-elif [ -n "$TRAVIS_PULL_REQUEST_BRANCH" ]; then
-  export IS_RUNNING_ON_TRAVIS=true
-  HATCHET_BUILDPACK_BRANCH="$TRAVIS_PULL_REQUEST_BRANCH"
+elif [ -n "$ci_branch" ]; then
+  HATCHET_BUILDPACK_BRANCH="$ci_branch"
 else
   HATCHET_BUILDPACK_BRANCH=$(git name-rev HEAD 2> /dev/null | sed 's#HEAD\ \(.*\)#\1#' | sed 's#tags\/##')
 fi