+++ lib/binaries.sh
@@ -1,45 +1,23 @@
 #!/usr/bin/env bash
 
+# Compiled from: https://github.com/heroku/buildpacks-nodejs/blob/main/common/nodejs-utils/src/bin/resolve_version.rs
 RESOLVE="$BP_DIR/lib/vendor/resolve-version-$(get_os)"
-RESOLVE_V2="$BP_DIR/lib/vendor/resolve"
 
 resolve() {
   local binary="$1"
   local versionRequirement="$2"
-  local n=0
-  local output v2_output resolve_is_equal
+  local output
 
-  # retry this up to 5 times in case of spurious failed API requests
-  until [ $n -ge 5 ]
-  do
-    # if a user sets the HTTP_PROXY ENV var, it could prevent this from making the S3 requests
-    # it needs here. We can ignore this proxy for aws urls with NO_PROXY. Some environments
-    # require a proxy for all HTTP requests, so the NO_PROXY ENV var should be set outside the
-    # script by the user
-    # see testAvoidHttpProxyVersionResolutionIssue test and README
-    if output=$($RESOLVE "$binary" "$versionRequirement"); then
-      v2_output=$($RESOLVE_V2 "$BP_DIR/inventory/$binary.toml" "$versionRequirement")
-      resolve_is_equal=$(if [[ "$output" == "$v2_output" ]]; then echo true; else echo false; fi)
-
-      meta_set "resolve-v1-$binary" "$output"
-      meta_set "resolve-v2-$binary" "$v2_output"
-      meta_set "resolve-is-equal-$binary" "$resolve_is_equal"
-      meta_set "resolve-v2-error" "$STD_ERR"
-
-      echo "$output"
-      return 0
-    # don't retry if we get a negative result
-    elif [[ $output = "No result" ]]; then
-      return 1
-    elif [[ $output == "Could not parse"* ]] || [[ $output == "Could not get"* ]]; then
+  if output=$($RESOLVE "$BP_DIR/inventory/$binary.toml" "$versionRequirement"); then
+    meta_set "resolve-v2-$binary" "$output"
+    meta_set "resolve-v2-error" "$STD_ERR"
+    if [[ $output = "No result" ]]; then
       return 1
     else
-      n=$((n+1))
-      # break for a second with a linear backoff
-      sleep $((n+1))
+      echo $output
+      return 0
     fi
-  done
-
+  fi
   return 1
 }
 
@@ -64,7 +42,7 @@ install_yarn() {
     echo "Downloading and installing yarn ($number)"
   fi
 
-  code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 -o /tmp/yarn.tar.gz --write-out "%{http_code}")
+  code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 --retry-connrefused --connect-timeout 5 -o /tmp/yarn.tar.gz --write-out "%{http_code}")
 
   if [ "$code" != "200" ]; then
     echo "Unable to download yarn: $code" && false
@@ -79,6 +57,8 @@ install_yarn() {
   fi
   chmod +x "$dir"/bin/*
 
+  # Verify yarn works before capturing and ensure its stderr is inspectable later
+  yarn --version 2>&1 1>/dev/null
   if $YARN_2; then
     echo "Using yarn $(yarn --version)"
   else
@@ -87,9 +67,20 @@ install_yarn() {
 }
 
 install_nodejs() {
-  local version=${1:-16.x}
+  local version="${1:-}"
   local dir="${2:?}"
   local code resolve_result
+
+  if [[ -z "$version" ]]; then
+    # Node.js 18+ is incompatible with ubuntu:18 (and thus heroku-18) because of a libc mismatch:
+    # node: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.28' not found (required by node)
+    # Fallback to a 16.x default for heroku-18 until heroku-18 or Node.js 16.x are EOL.
+    if [[ "$STACK" == "heroku-18" ]]; then
+      version="16.x"
+    else
+      version="18.x"
+    fi
+  fi
 
   if [[ -n "$NODE_BINARY_URL" ]]; then
     url="$NODE_BINARY_URL"
@@ -107,7 +98,7 @@ install_nodejs() {
     echo "Downloading and installing node $number..."
   fi
 
-  code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz --write-out "%{http_code}")
+  code=$(curl "$url" -L --silent --fail --retry 5 --retry-max-time 15 --retry-connrefused --connect-timeout 5 -o /tmp/node.tar.gz --write-out "%{http_code}")
 
   if [ "$code" != "200" ]; then
     echo "Unable to download node: $code" && false
@@ -122,6 +113,8 @@ install_npm() {
   local version="$1"
   local dir="$2"
   local npm_lock="$3"
+  # Verify npm works before capturing and ensure its stderr is inspectable later
+  npm --version 2>&1 1>/dev/null
   npm_version="$(npm --version)"
 
   # If the user has not specified a version of npm, but has an npm lockfile
@@ -140,6 +133,8 @@ install_npm() {
     if ! npm install --unsafe-perm --quiet -g "npm@$version" 2>@1>/dev/null; then
       echo "Unable to install npm $version; does it exist?" && false
     fi
+    # Verify npm works before capturing and ensure its stderr is inspectable later
+    npm --version 2>&1 1>/dev/null
     echo "npm $(npm --version) installed"
   fi
 }