+++ spec/ci/node_18_spec.rb
@@ -1,0 +1,27 @@
+require_relative '../spec_helper'
+
+describe "Hello World for Node v18.x" do
+  context "a single-process Node v18.x app" do
+    let(:app) {
+      Hatchet::Runner.new("spec/fixtures/repos/node-18")
+    }
+
+    it "should deploy successfully" do
+      app.deploy do |app|
+        expect(successful_body(app).strip).to eq("Hello, world!")
+      end
+    end
+  end
+
+  context "on an incompatible stack (heroku-18)" do
+    it "should log a stack compatibility message" do
+      Hatchet::Runner.new(
+        "spec/fixtures/repos/node-18",
+        stack: "heroku-18",
+        allow_failure: true
+      ).deploy do |app|
+        expect(app.output).to match("Node.js version is not compatible with the current stack")
+      end
+    end
+  end
+end