+++ spec/spec_helper.rb
@@ -32,7 +32,7 @@ def successful_body(app, options = {})
               retry_interval: 0.5,
               retry_limit:    retry_limit
            ).body
-rescue Excon::HTTPStatus => e
+rescue Excon::Error::HTTPStatus => e
   puts e.response.body
   raise e
 end
@@ -77,7 +77,7 @@ def resolve_all_supported_node_versions(options = {})
 end
 
 def version_supports_metrics(version)
-  SemVersion.new(version).satisfies?('>= 10.0.0') && SemVersion.new(version).satisfies?('< 16.0.0')
+  SemVersion.new(version).satisfies?('>= 10.0.0') && SemVersion.new(version).satisfies?('< 20.0.0')
 end
 
 def get_test_versions
@@ -86,7 +86,7 @@ def get_test_versions
   elsif ENV['TEST_ALL_NODE_VERSIONS'] == 'true'
     versions = resolve_all_supported_node_versions()
   else
-    versions = resolve_node_version(['12.x', '14.x', '16.x', '17.x'])
+    versions = resolve_node_version(['14.x', '16.x', '18.x', '19.x'])
   end
   puts("Running tests for Node versions: #{versions.join(', ')}")
   versions